requirements_streamlit.txt
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
torch==2.0.1
torchvision==0.15.2
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
scikit-image==0.22.0
python-multipart==0.0.6
